feat: Add slots for video and file upload components and alerts
This change add plugin slots for the file and video upload components, and the alerts components on those pages.

(cherry picked from commit 9c79a4184a95ecc6ac06ce8b56c64227af7528ef)
