+++ src/plugin-slots/CourseOutlinePageAlertsSlot/index.jsx
@@ -1,0 +1,5 @@
+import { PluginSlot } from '@openedx/frontend-plugin-framework';
+import React from 'react';
+
+const CourseOutlinePageAlertsSlot = () => <PluginSlot id="course_outline_page_alerts_slot" />;
+export default CourseOutlinePageAlertsSlot;
